Move cosigner-shares command into package, Dockerfile for horcrux
