+++ Makefile
@@ -1,17 +1,12 @@
-GOLINT:=$(shell go list -f {{.Target}} golang.org/x/lint/golint)
-
 all: build
 
-build: build/signer build/key2shares
+build: build/signer
 
 build/signer: cmd/signer/main.go $(wildcard internal/**/*.go)
 	CGO_ENABLED=0 go build -o ./build/signer ${gobuild_flags} ./cmd/signer
 
-build/key2shares: cmd/key2shares/main.go $(wildcard internal/**/*.go)
-	CGO_ENABLED=0 go build -o ./build/key2shares ${gobuild_flags} ./cmd/key2shares
-
-lint: tools
-	@$(GOLINT) -set_exit_status ./...
+build-linux:
+	go build -o ./build/signer ./cmd/signer
 
 test:
 	@go test -short ./...
@@ -32,3 +27,6 @@ clean:
 
 build-simd-docker:
 	docker build -t jackzampolin/simd:v0.42.3 -f ./docker/simd/Dockerfile ./docker/simd/
+
+build-horcrux-docker:
+	docker build -t jackzampolin/horcrux:v0.1.0 -f ./docker/horcrux/Dockerfile .
