+++ docker/horcrux/Dockerfile
@@ -1,0 +1,21 @@
+FROM golang:1.16-alpine AS build-env
+
+ENV PACKAGES make
+
+RUN apk add --no-cache $PACKAGES
+
+WORKDIR /go/src/github.com/jackzampolin/horcrux
+
+ADD . .
+
+RUN make build-linux
+
+FROM alpine:edge
+
+RUN apk add --no-cache ca-certificates
+
+WORKDIR /root
+
+COPY --from=build-env /go/src/github.com/jackzampolin/horcrux/build/signer /usr/bin/signer
+
+CMD ["signer"]
